requirements.txt
uvicorn==0.27.1
fastapi==0.110.0
fastapi-jwt-auth==0.5.0
pydantic==1.10.1
pydantic_core==2.10.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9


python-multipart==0.0.9
python-dotenv==0.21.0
isort==5.13.2
bcrypt==4.1.2
